+++ CHANGELOG.md
@@ -9,7 +9,8 @@ All notable changes to this project will be documented in this file.
 - Upgrade vite to v3
 - Improve about page modules
 - Improve sponsor style
-- Replace `sanitize-html` with `dompurify` [#9200](https://github.com/vitejs/vite/issues/9200) [#560](https://github.com/apostrophecms/sanitize-html/issues/560)
+- Improve `mapbox-gl`
+- remove `sanitize-html` and `dompurify` [#9200](https://github.com/vitejs/vite/issues/9200) [#560](https://github.com/apostrophecms/sanitize-html/issues/560)
 
 **Fix**